Add beginning of parser code
To make the code a little bit more flexible the user should be allowed
to pass in the various magic numbers.
The user will be able to change: the notification timeout; when the tool
should start shutting down the computer from; and when the tool should
stop trying to shutdown the computer.
The current magic numbers will be used as defaults.